backend/requirements.txt
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-colorfield==0.7.2
requests==2.26.0
pytest-django==4.4.0
pytest==6.2.4
pytest-pythonpath==0.7.3
asgiref==3.2.10
djoser==2.1.0
drf_extra_fields==3.4.0
gunicorn==20.0.4
Pillow==9.2.0
psycopg2-binary
python-dotenv==0.20.0
pytz==2020.1
requests==2.26.0
reportlab==3.6.12
sqlparse==0.3.1